requirements.txt
setuptools
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
python-decouple==3.8
weasyprint==62.3
reportlab==4.1.0
cashfree-sdk==0.1.3
django-widget-tweaks==1.4.12
dj-database-url==2.1.0